// src/recipe/thirdparty/types.ts
export interface TypeProviderGetResponse {
  accessTokenAPI: {
    url: string
    params: { [key: string]: string } // Will be merged with our object
  }
  authorisationRedirect: {
    url: string
    params: { [key: string]: string | ((request: any) => string) }
  }
  getProfileInfo: (authCodeResponse: any, userContext: any) => Promise<UserInfo>
  getClientId: (userContext: any) => string
  getRedirectURI?: (userContext: any) => string // if undefined, the redirect_uri is set on the frontend.
}

export interface TypeProvider {
  id: string
  get: (
    redirectURI: string | undefined,
    authCodeFromRequest: string | undefined,
    userContext: any
  ) => TypeProviderGetResponse
  isDefault?: boolean // if not present, we treat it as false
}

export interface User {
  // https://github.com/supertokens/core-driver-interface/wiki#third-party-user
  id: string
  timeJoined: number
  email: string
  thirdParty: {
    id: string
    userId: string
  }
}

export interface TypeInputSignInAndUp {
  providers: TypeProvider[]
}

export interface TypeNormalisedInputSignInAndUp {
  providers: TypeProvider[]
}

export interface TypeInput {
  signInAndUpFeature: TypeInputSignInAndUp
  override?: {
    functions?: (
      originalImplementation: RecipeInterface,
      builder?: OverrideableBuilder<RecipeInterface>
    ) => RecipeInterface
    apis?: (originalImplementation: APIInterface, builder?: OverrideableBuilder<APIInterface>) => APIInterface
  }
}

export interface TypeNormalisedInput {
  signInAndUpFeature: TypeNormalisedInputSignInAndUp
  override: {
    functions: (
      originalImplementation: RecipeInterface,
      builder?: OverrideableBuilder<RecipeInterface>
    ) => RecipeInterface
    apis: (originalImplementation: APIInterface, builder?: OverrideableBuilder<APIInterface>) => APIInterface
  }
}

export interface RecipeInterface {
  getUserById(input: { userId: string; userContext: any }): Promise<User | undefined>

  getUsersByEmail(input: { email: string; userContext: any }): Promise<User[]>

  getUserByThirdPartyInfo(input: {
    thirdPartyId: string
    thirdPartyUserId: string
    userContext: any
  }): Promise<User | undefined>

  signInUp(input: {
    thirdPartyId: string
    thirdPartyUserId: string
    email: string
    userContext: any
  }): Promise<{ status: 'OK'; createdNewUser: boolean; user: User }>
}

export interface APIOptions {
  recipeImplementation: RecipeInterface
  config: TypeNormalisedInput
  recipeId: string
  isInServerlessEnv: boolean
  providers: TypeProvider[]
  req: BaseRequest
  res: BaseResponse
  appInfo: NormalisedAppinfo
}

export interface APIInterface {
  authorisationUrlGET:
  | undefined
  | ((input: {
    provider: TypeProvider
    options: APIOptions
    userContext: any
  }) => Promise<
              | {
                status: 'OK'
                url: string
              }
              | GeneralErrorResponse
          >)

  signInUpPOST:
  | undefined
  | ((input: {
    provider: TypeProvider
    code: string
    redirectURI: string
    authCodeResponse?: any
    clientId?: string
    options: APIOptions
    userContext: any
  }) => Promise<
              | {
                status: 'OK'
                createdNewUser: boolean
                user: User
                session: SessionContainerInterface
                authCodeResponse: any
              }
              | { status: 'NO_EMAIL_GIVEN_BY_PROVIDER' }
              | GeneralErrorResponse
          >)

  appleRedirectHandlerPOST:
  | undefined
  | ((input: { code: string; state: string; options: APIOptions; userContext: any }) => Promise<void>)

}